// lib/gsc/types.ts
export interface GSCMetrics {
  clicks: number
  impressions: number
  ctr: number
  position: number
}

export interface GSCTimeSeriesData {
  date: string
  clicks: number
  impressions: number
  ctr: number
  position: number
}

export interface GSCPageData {
  page: string
  clicks: number
  impressions: number
  ctr: number
  position: number
}